Create class op sparse
